+++ src/components/Quiz/ProgressBar.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import './main.css'
 
 export default function ProgressBar(props) {
     return (
