+++ src/components/Quiz/Quiz.jsx
@@ -1,6 +1,6 @@
 import React, {useState} from 'react';
 import {questions} from './questions';
-import ProgressBar from '../ProgressBar';
+import ProgressBar from './ProgressBar';
 
 export default function Quiz(){
     const [displayedQuestion, setDisplayedQuestion] = useState(0);
@@ -30,8 +30,8 @@ export default function Quiz(){
                 <div className='question-section'>
                     <div className='question-count'>Question {displayedQuestion +1} / {questions.length}</div>
                     <div className='question-text'>{questions[displayedQuestion].questionText}</div>
-                    <div className='question-answers'>
-                        {questions[displayedQuestion].answerOptions.map((answerOption) => <button onClick={() => handleAnswerButton(answerOption.isCorrect)} key={answerOption.key}>{answerOption.answerText}</button>)}
+                    <div className='answer-section'>
+                        {questions[displayedQuestion].answerOptions.map((answerOption) => <button className = 'answer-button' onClick={() => handleAnswerButton(answerOption.isCorrect)} key={answerOption.key}>{answerOption.answerText}</button>)}
                     </div>
                 </div>
             </>)}
