+++ src/main.css
@@ -1,0 +1,84 @@
+:root {
+  --purple: #a64ac9;
+  --yellow: #fccd04;
+  --blue: #202a52;
+}
+
+body {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
+    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
+    sans-serif;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  min-height: 100vh;
+}
+
+/* Quiz  */
+.question-card {
+  background-color: var(--purple);
+  width: 550px;
+  min-height: 300px;
+  border-radius: 15px;
+  padding: 20px;
+  box-shadow: 10px 10px 42px 0px rgba(0, 0, 0, 0.75);
+  justify-content: space-evenly;
+}
+
+/* questions  */
+.question-section {
+  width: 100%;
+  position: relative;
+}
+
+.question-count {
+  margin-bottom: 5px;
+}
+
+.question-text {
+  margin-bottom: 12px;
+}
+.answer-section {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+/* answer */
+.answer-button {
+  width: 100%;
+  font-size: 16px;
+  color: #ffffff;
+  background-color: var(--blue);
+  border-color: var(--blue);
+  border-radius: 20px;
+  margin-bottom: 15px;
+  display: flex;
+  padding: 10px;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+}
+
+button:hover {
+  background-color: #555e7d;
+}
+
+/* progress bar component */
+.progress-bar {
+  position: relative;
+  height: 20px;
+  width: 350px;
+  border-radius: 50px;
+  border: 1px solid #333;
+  margin: 0px auto 5px auto;
+}
+
+.filler {
+  background: var(--yellow);
+  height: 100%;
+  border-radius: inherit;
+  transition: width 0.2s ease-in;
+}
